# spec/openapi.yaml
openapi: 3.0.1
info:
  description: An API suite providing read only access to search and retrieve
    public company data
  title: Companies House Public Data API
  version: 1.0.0
servers:
  - url: https://api.company-information.service.gov.uk/
  - url: http://api.company-information.service.gov.uk/
security:
  - api_key: []
...
paths:
...
  /alphabetic-search/companies:
    get:
      description: Search for a company
      operationId: alphabetic_searchCompaniesGET
      parameters:
        - description: The company name being searched for
          in: query
          name: q
          required: true
          schema:
            type: string
        - description: The ordered_alpha_key_with_id used for paging
          in: query
          name: search_above
          schema:
            type: string
        - description: The ordered_alpha_key_with_id used for paging
          in: query
          name: search_below
          schema:
            type: string
        - description: The maximum number of results matching the search term(s) to return
            with a range of 1 to 100
          in: query
          name: size
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                title: List of companies
                type: object
          description: A list of companies
        "404":
          content: {}
          description: No companies found
        "422":
          content: {}
          description: Invalid size parameter, size must be greater than zero and not
            greater than 100
      summary: Search for a company
      tags:
        - search
      x-operationName: search companies alphabetically
      x-eov-operation-handler: controllers/SearchController
  /advanced-search/companies:
    get:
      description: Advanced search for a company
      operationId: advanced_searchCompaniesGET
      parameters:
        - description: The company name (must contain) advanced search filter
          in: query
          name: company_name
          schema:
            type: string
        - description: The company status advanced search filter. To search using multiple
            values, use a comma delimited list or multiple of the same key i.e.
            company_status=xxx&company_status=yyy
          in: query
          name: company_status
          schema: {}
        - description: The company subtype advanced search filter. To search using
            multiple values, use a comma delimited list or multiple of the same
            key i.e. company_subtype=xxx&company_subtype=yyy
          in: query
          name: company_subtype
          schema:
            type: string
        - description: The company type advanced search filter. To search using multiple
            values, use a comma delimited list or multiple of the same key i.e.
            company_type=xxx&company_type=yyy
          in: query
          name: company_type
          schema: {}
        - description: The dissolved from date advanced search filter
          in: query
          name: dissolved_from
          schema: {}
        - description: The dissolved to date advanced search filter
          in: query
          name: dissolved_to
          schema: {}
        - description: The incorporated from date advanced search filter
          in: query
          name: incorporated_from
          schema: {}
        - description: The incorporated to date advanced search filter
          in: query
          name: incorporated_to
          schema: {}
        - description: The location advanced search filter
          in: query
          name: location
          schema:
            type: string
        - description: The SIC codes advanced search filter. To search using multiple
            values, use a comma delimited list or multiple of the same key i.e.
            sic_codes=xxx&sic_codes=yyy
          in: query
          name: sic_codes
          schema: {}
        - description: The maximum number of results matching the search term(s) to return
            with a range of 1 to 5000
          in: query
          name: size
          schema:
            type: string
        - description: The point at which results will start from i.e show search results
            from result 20 (used for paging)
          in: query
          name: start_index
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                title: A list of companies
                type: object
          description: A list of companies
        "400":
          content: {}
          description: Bad request
        "404":
          content: {}
          description: No companies found
      summary: Advanced search for a company
      tags:
        - search
      x-operationName: advanced company search
      x-eov-operation-handler: controllers/SearchController
...
components:
...
  securitySchemes:
    api_key:
      in: header
      name: api_key
      type: apiKey
...